Создает класс родитель. Выносит вспомогательные функции в модули.
