+++ src/component.js
@@ -1,0 +1,35 @@
+import createElement from './create-element';
+
+export default class Component {
+  constructor() {
+    if (new.target === Component) {
+      throw new Error(`Can't instantiate Component, only concrete one.`);
+    }
+  }
+
+  get template() {
+    throw new Error(`You have to define template.`);
+  }
+
+  get element() {
+    return this._element;
+  }
+
+  bind() {
+  }
+
+  unbind() {
+  }
+
+  render() {
+    this._element = createElement(this.template);
+    this.bind();
+    return this._element;
+  }
+
+  unrender() {
+    this.unbind();
+    this._element.remove();
+    this._element = null;
+  }
+}
