+++ src/point.js
@@ -1,7 +1,10 @@
-import createElement from '../src/create-element';
+import Component from './component';
+import getTime from './get-time';
+import getDuration from './get-duration';
 
-export default class Point {
+export default class Point extends Component {
   constructor(data) {
+    super();
     this._price = data.price;
     this._type = data.type;
     this._destination = data.destination;
@@ -17,24 +20,6 @@ export default class Point {
     if (typeof this._onEdit === `function`) {
       this._onEdit();
     }
-  }
-
-  _getTime(millisec) {
-    let minutes = new Date(millisec).getMinutes();
-    let hours = new Date(millisec).getHours();
-
-    hours = (hours >= 10) ? hours : `0` + hours;
-    minutes = (minutes >= 10) ? minutes : `0` + minutes;
-
-    return `${hours}:${minutes}`;
-  }
-
-  _getDuration(startTime, endTime) {
-    const duration = endTime - startTime;
-    const minutes = new Date(duration).getMinutes();
-    const hours = new Date(duration).getHours();
-
-    return `${hours}h ${minutes}m`;
   }
 
   _getOffers(offersArray) {
@@ -54,17 +39,13 @@ export default class Point {
             </li>`).join(``);
   }
 
-  get element() {
-    return this._element;
-  }
-
   get template() {
     return `<article class="trip-point">
           <i class="trip-icon">${this._type.icon}</i>
           <h3 class="trip-point__title">${this._type.name} to ${[...this._destination][Math.floor(Math.random() * 5)]}</h3>
           <p class="trip-point__schedule">
-            <span class="trip-point__timetable">${this._getTime(this._time.start)}&nbsp;&mdash; ${this._getTime(this._time.end)}</span>
-            <span class="trip-point__duration">${this._getDuration(this._time.start, this._time.end)}</span>
+            <span class="trip-point__timetable">${getTime(this._time.start)}&nbsp;&mdash; ${getTime(this._time.end)}</span>
+            <span class="trip-point__duration">${getDuration(this._time.start, this._time.end)}</span>
           </p>
           <p class="trip-point__price">&euro;&nbsp;${this._price}</p>
           <ul class="trip-point__offers">
@@ -85,17 +66,5 @@ export default class Point {
   unbind() {
     this._element.closest(`.trip-point`)
       .removeEventListener(`click`, this._onPointClick);
-  }
-
-  render() {
-    this._element = createElement(this.template);
-    this.bind();
-    return this._element;
-  }
-
-  unrender() {
-    this.unbind();
-    this._element.remove();
-    this._element = null;
   }
 }
